// src/anomalies/enums/anomaly.enum.ts
export enum AnomalyType {
  REAL_ANOMALY = 'REAL_ANOMALY',
  EXPLAINABLE_ANOMALY = 'EXPLAINABLE_ANOMALY',
  FALSE_POSITIVE = 'FALSE_POSITIVE',
  DATA_QUALITY = 'DATA_QUALITY',
  PENDING_ANALYSIS = 'PENDING_ANALYSIS',
}

export enum AnomalySeverity {
  LOW = 'LOW',
  MEDIUM = 'MEDIUM',
  HIGH = 'HIGH',
  PENDING = 'PENDING',
}

export enum AnomalyStatus {
  ANALYZING = 'ANALYZING',
  COMPLETED = 'COMPLETED',
  FAILED = 'FAILED',
  RESOLVED = 'RESOLVED',
  DETECTED = 'DETECTED',
}